CarDAOImpl замена Session на Query

+++ src/main/java/com/github/elizalex/garage/entity/Car.java
@@ -1,9 +1,6 @@
 package com.github.elizalex.garage.entity;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 import javax.persistence.*;
 import java.util.List;
@@ -16,7 +13,7 @@ public class Car {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "car_id")
-    private Long id;
+    private int id;
 
     @Column(name = "car_brand")
     private String brand;
